+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Add tests
+- Add `EnvironmentDefinition` Type and Module
 
 ## 1.0.0 - 2020-02-10
 - Initial implementation